+++ internal/checks/self_write_identity_other.go
@@ -1,0 +1,9 @@
+//go:build !darwin && !linux
+
+package checks
+
+import "os"
+
+func selfWriteIdentityFromFileInfo(os.FileInfo) (selfWriteFileIdentity, bool) {
+	return selfWriteFileIdentity{}, false
+}